Support anon subs better
Possibly the patch is self-evident, but the idea is to detect when
an anon sub is being declared, and stick a placeholder entry in @Subs.
These are distinguished by having name => undef.

That way, we don't have to change any of the code that accumulates sub
metadata (for example parameters). We can accumulate it, same as always,
but then filter out anon subs before generating our documentation.

A previous iteration tried to make metadata accumulation smarter, by
checking whether @Subs had any entries yet (since the original error was
trying to access @Subs[-1] when @Subs was length 0). But this wasn't
quite smart enough - it only handled cases where no named subs had been
declared yet. If a named sub *had* been declared, it would get all the
metadata for the anon subs following it, which is clearly wrong! The
easiest and most logical fix is to acknowledge anon subs, but throw them
away later.